Move fireworks-specific API types into types.go
